requirements.txt
absl-py==1.2.0
astunparse==1.6.3
async-timeout==4.0.2
cachetools==5.2.0
certifi==2022.9.14
charset-normalizer==2.1.1
click==8.1.3
Deprecated==1.2.13
Flask==2.2.2
flatbuffers==22.9.24
gast==0.4.0
google-auth==2.13.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.49.1
h5py==3.7.0
idna==3.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
keras==2.10.0
Keras-Preprocessing==1.1.2
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
numpy==1.23.4
oauthlib==3.2.1
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.4
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.4
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.1.2
scipy==1.9.2
six==1.16.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
termcolor==2.0.1
threadpoolctl==3.1.0
typing-extensions==4.4.0
urllib3==1.26.12
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.9.0
python-dotenv==0.21.0
boto3~=1.26.22
redis[hiredis]==4.4.2
pylogger-discord==0.0.122
